Preemption logic.
When spin_lock() is called it increments the preemption
counter of the task. In that case (unless it already was
disabled) we have to raise Windows IRQL (of the processor
where the task is running) to DISPATCH_LEVEL in order
to prevent Windows from re-scheduling the task. On the
other hand when spin_unlock() is called it decrements
the preemption counter, in which case we have to
enable Windows preemption again (if the preemption counter
reached 0).