Working on memory.
Reason for problems seems to be that the memblock is used
for early alloc, but also for figuring out total memory
and its location. For early alloc it needs to be a valid
Windows virtual address of allocated memory (such as 0xb6761000)
but for total memory it should start at 0. Need to hack at
certain places.